requirements.txt
langgraph
langchain
langchain-openai
langchain_community
tavily-python
python-dotenv
gradio
faiss-cpu